Move mocha config to own file
