+++ truffle-config.js
@@ -1,35 +1,4 @@
-
-// Thanks AragonOS for this conditional reporter logic.
-const mochaGasSettingsShell = {
-  reporter: 'eth-gas-reporter',
-  reporterOptions : {
-    onlyCalledMethods: true,
-  }
-}
-
-const mochaGasSettingsDocs = {
-  reporter: 'eth-gas-reporter',
-  reporterOptions : {
-    onlyCalledMethods: true,
-    rst: true,
-    rstTitle: 'Appendix: Gas Usage',
-    outputFile: 'docs/Gas.rst',
-    noColors: true
-  }
-}
-
-const mochaGeth = {
-  grep: "geth",
-}
-
-let mocha = {};
-if (process.env.GAS_REPORTER){
-  mocha = mochaGasSettingsShell
-} else if (process.env.GAS_DOCS){
-  mocha = mochaGasSettingsDocs
-} else if (process.env.NETWORK === 'geth'){
-  mocha = mochaGeth
-}
+const mocha = require('./scripts/mochaSettings');
 
 module.exports = {
   networks: {
